(reg_basic_block): Now int *, not short *.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@4173 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/basic-block.h
@@ -65,4 +65,4 @@ extern regset *basic_block_live_at_start;
 
 #define REG_BLOCK_UNKNOWN -1
 #define REG_BLOCK_GLOBAL -2
-extern short *reg_basic_block;
+extern int *reg_basic_block;
